requirements.txt
flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Werkzeug>=3.0.0
Pillow>=10.0.0
email-validator==2.0.0